make printing the responsibility of a PrintStrategy
